requirements.txt
absl-py==2.1.0
accelerate==0.30.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
apyori==1.1.2
astunparse==1.6.3
async-timeout==4.0.3
blinker==1.8.2
blis==0.7.11
brotlipy==0.7.0
cachetools==5.3.3
catalogue==2.0.10
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.2.2
conda==22.11.1
confection==0.1.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.8
datasets==2.19.1
dill==0.3.8
evaluate==0.4.2
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.38.0
frozenlist==1.4.1
fsspec==2024.3.1
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.64.0
h5py==3.11.0
huggingface-hub==0.23.1
imbalanced-learn==0.12.2
imblearn==0.0
intel-openmp==2021.4.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
Jinja2==3.1.4
joblib==1.3.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyterlab_widgets==3.0.10
kaggle==1.6.14
keras==3.3.3
kiwisolver==1.4.4
langcodes==3.3.0
libclang==18.1.1
mahotas==1.4.13
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.0
mdurl==0.1.2
mkl==2021.4.0
mkl-service==2.4.0
ml-dtypes==0.3.2
mlxtend==0.23.1
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mysql-connector-python==8.0.32
namex==0.0.8
networkx==3.3
nltk==3.8.1
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
pandas==2.2.1
patsy==0.5.6
Pillow==9.4.0
preshed==3.0.9
protobuf==4.25.3
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.9.1
pydotplus==2.0.2
pyparsing==3.0.9
pytesseract==0.3.10
python-slugify==8.0.4
pywin32==305.1
PyYAML==6.0.1
qtconsole==5.5.2
QtPy==2.4.1
regex==2023.12.25
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.11.4
seaborn==0.13.2
smart-open==6.4.0
smmap==5.0.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
statsmodels==0.14.1
streamlit==1.35.0
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
tf_keras==2.16.0
thinc==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tokenizers==0.19.1
toml==0.10.2
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
transformers==4.41.1
typer==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
ultralytics==8.2.22
wasabi==1.1.2
watchdog==4.0.1
weasel==0.3.4
webencodings==0.5.1
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wincertstore==0.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4